+++ spec/lib/sendgrid_actionmailer_spec.rb
@@ -3,6 +3,10 @@ require 'webmock/rspec'
 
 module SendGridActionMailer
   describe DeliveryMethod do
+    def transform_keys(object, &block)
+      SendGridActionMailer::DeliveryMethod.transform_keys(object, &block)
+    end
+
     subject(:mailer) do
       DeliveryMethod.new(api_key: 'key')
     end
@@ -36,6 +40,11 @@ module SendGridActionMailer
         expect(m.settings[:api_key]).to eq('ABCDEFG')
       end
 
+      it 'has correct host' do
+        m = DeliveryMethod.new(host: 'example.com')
+        expect(m.settings[:host]).to eq('example.com')
+      end
+
       it 'default raise_delivery_errors' do
         m = DeliveryMethod.new()
         expect(m.settings[:raise_delivery_errors]).to eq(false)
@@ -54,6 +63,16 @@ module SendGridActionMailer
       it 'sets return_response' do
         m = DeliveryMethod.new(return_response: true)
         expect(m.settings[:return_response]).to eq(true)
+      end
+
+      it 'sets perform_deliveries' do
+        m = DeliveryMethod.new(perform_send_request: false)
+        expect(m.settings[:perform_send_request]).to eq(false)
+      end
+
+      it 'sets http_options' do
+        m = DeliveryMethod.new(http_options: {open_timeout: 40})
+        expect(m.settings[:http_options]).to eq({open_timeout: 40})
       end
     end
 
@@ -97,11 +116,11 @@ module SendGridActionMailer
         end
 
         it 'sets dynamic api_key, but should revert to default settings api_key' do
-          expect(SendGrid::API).to receive(:new).with(api_key: 'key')
+          expect(SendGrid::API).to receive(:new).with(api_key: 'key', http_options: {})
           mailer.deliver!(default)
-          expect(SendGrid::API).to receive(:new).with(api_key: 'test_key')
+          expect(SendGrid::API).to receive(:new).with(api_key: 'test_key', http_options: {})
           mailer.deliver!(mail)
-          expect(SendGrid::API).to receive(:new).with(api_key: 'key')
+          expect(SendGrid::API).to receive(:new).with(api_key: 'key', http_options: {})
           mailer.deliver!(default)
         end
       end
@@ -242,12 +261,50 @@ module SendGridActionMailer
         ])
       end
 
-      context 'send options' do
-        it 'sets a template_id' do
+      context 'template_id' do
+        before do
           mail['template_id'] = '1'
+        end
+
+        it 'sets a template_id' do
           mailer.deliver!(mail)
           expect(client.sent_mail['template_id']).to eq('1')
         end
+
+        it 'does not set unsubscribe substitutions' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'].first).to_not have_key('substitutions')
+        end
+
+        it 'does not set send a content type' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['content']).to eq(nil)
+        end
+
+        it 'does not set send a content type even if body is given' do
+          # This matches the default behavior of ActionMail. body must be
+          # specified and content_type defaults to text/plain.
+          mail.body = 'I heard you like pineapple.'
+          mail.content_type = 'text/plain'
+          mailer.deliver!(mail)
+          expect(client.sent_mail['content']).to eq(nil)
+        end
+      end
+
+      context 'without dynamic template data or a template id' do
+        it 'sets unsubscribe substitutions' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'].first).to have_key('substitutions')
+          substitutions = client.sent_mail['personalizations'].first['substitutions']
+          expect(substitutions).to eq({
+            '%asm_group_unsubscribe_raw_url%' => '<%asm_group_unsubscribe_raw_url%>',
+            '%asm_global_unsubscribe_raw_url%' => '<%asm_global_unsubscribe_raw_url%>',
+            '%asm_preferences_raw_url%' => '<%asm_preferences_raw_url%>'
+          })
+        end
+      end
+
+      context 'send options' do
 
         it 'sets sections' do
           mail['sections'] = {'%foo%' => 'bar'}
@@ -283,10 +340,10 @@ module SendGridActionMailer
         end
 
         it 'sets asm' do
-          asm = {'group_id' => 99, 'groups_to_display' => [4,5,6,7,8]}
+          asm = {group_id: 99, groups_to_display: [4,5,6,7,8]}
           mail['asm'] = asm
           mailer.deliver!(mail)
-          expect(client.sent_mail['asm']).to eq(asm)
+          expect(client.sent_mail['asm']).to eq(transform_keys(asm, &:to_s))
         end
 
         it 'sets ip_pool_name' do
@@ -295,109 +352,109 @@ module SendGridActionMailer
           expect(client.sent_mail['ip_pool_name']).to eq('marketing')
         end
 
-        context 'parse object' do
-          it "should parse 1.8 hash" do
-            asm = {'group_id' => 99, 'groups_to_display' => [4,5,6,7,8]}
-            mail['asm'] = asm
-            mailer.deliver!(mail)
-            expect(client.sent_mail['asm']).to eq({"group_id" => 99, "groups_to_display" => [4,5,6,7,8]})
-          end
-
-          it "should parse 1.9 hash" do
-            asm = { group_id: 99, groups_to_display: [4,5,6,7,8]}
-            mail['asm'] = asm
-            mailer.deliver!(mail)
-            expect(client.sent_mail['asm']).to eq({"group_id" => 99, "groups_to_display" => [4,5,6,7,8]})
-          end
-
-          it "should parse json" do
-            asm = {'group_id' => 99, 'groups_to_display' => [4,5,6,7,8]}
-            mail['asm'] = asm.to_json
-            mailer.deliver!(mail)
-            expect(client.sent_mail['asm']).to eq({"group_id" => 99, "groups_to_display" => [4,5,6,7,8]})
-          end
+        it 'should not change values inside custom args' do
+          custom_args = { 'text' => 'line with a => in it' }
+          mail['custom_args'] = custom_args
+          mailer.deliver!(mail)
+          expect(client.sent_mail['custom_args']).to eq('text' => 'line with a => in it')
         end
 
         context 'mail_settings' do
           it 'sets bcc' do
-            bcc = { 'bcc' => { 'enable' => true, 'email' => 'test@example.com' }}
+            bcc = { bcc: { enable: true, email: 'test@example.com' }}
             mail['mail_settings'] = bcc
             mailer.deliver!(mail)
-            expect(client.sent_mail['mail_settings']).to eq(bcc)
+            expect(client.sent_mail['mail_settings']).to eq(transform_keys(bcc, &:to_s))
           end
 
           it 'sets bypass_list_management' do
-            bypass = { 'bypass_list_management' => { 'enable' => true }}
+            bypass = { bypass_list_management: { enable: true }}
             mail['mail_settings'] = bypass
             mailer.deliver!(mail)
-            expect(client.sent_mail['mail_settings']).to eq(bypass)
+            expect(client.sent_mail['mail_settings']).to eq(transform_keys(bypass, &:to_s))
           end
 
           it 'sets footer' do
-            footer = {'footer' => { 'enable' => true, 'text' => 'Footer Text', 'html' => '<html><body>Footer Text</body></html>'}}
+            footer = {footer: { enable: true, text: 'Footer Text', html: '<html><body>Footer Text</body></html>'}}
             mail['mail_settings'] = footer
             mailer.deliver!(mail)
-            expect(client.sent_mail['mail_settings']).to eq(footer)
+            expect(client.sent_mail['mail_settings']).to eq(transform_keys(footer, &:to_s))
           end
 
           it 'sets sandbox_mode' do
-            sandbox = {'sandbox_mode' => { 'enable' => true }}
+            sandbox = {sandbox_mode: { enable: true }}
             mail['mail_settings'] = sandbox
             mailer.deliver!(mail)
-            expect(client.sent_mail['mail_settings']).to eq(sandbox)
+            expect(client.sent_mail['mail_settings']).to eq(transform_keys(sandbox, &:to_s))
           end
 
           it 'sets spam_check' do
-            spam_check = {'spam_check' => { 'enable' => true, 'threshold' => 1, 'post_to_url' => 'https://spamcatcher.sendgrid.com'}}
+            spam_check = {spam_check: { enable: true, threshold: 1, post_to_url: 'https://spamcatcher.sendgrid.com'}}
             mail['mail_settings'] = spam_check
             mailer.deliver!(mail)
-            expect(client.sent_mail['mail_settings']).to eq(spam_check)
+            expect(client.sent_mail['mail_settings']).to eq(transform_keys(spam_check, &:to_s))
           end
         end
 
         context 'tracking_settings' do
           it 'sets click_tracking' do
-            tracking = { 'click_tracking' => { 'enable' => false, 'enable_text' => false }}
-            mail['tracking_settings'] = tracking
+            tracking = { click_tracking: { enable: false, enable_text: false }}
+            mail['tracking_settings'] = tracking.dup
             mailer.deliver!(mail)
-            expect(client.sent_mail['tracking_settings']).to eq(tracking)
+            expect(client.sent_mail['tracking_settings']).to eq(transform_keys(tracking, &:to_s))
           end
 
           it 'sets open_tracking' do
-            tracking = { 'open_tracking' => { 'enable' => true, 'substitution_tag' => 'Optional tag to replace with the open image in the body of the message' }}
+            tracking = { open_tracking: { enable: true, substitution_tag: 'Optional tag to replace with the open image in the body of the message' }}
             mail['tracking_settings'] = tracking
             mailer.deliver!(mail)
-            expect(client.sent_mail['tracking_settings']).to eq(tracking)
+            expect(client.sent_mail['tracking_settings']).to eq(transform_keys(tracking, &:to_s))
           end
 
           it 'sets subscription_tracking' do
-            tracking = { 'subscription_tracking' => { 'enable' => true, 'text' => 'text to insert into the text/plain portion of the message', 'html' => 'html to insert into the text/html portion of the message', 'substitution_tag' => 'Optional tag to replace with the open image in the body of the def message' }}
+            tracking = { subscription_tracking: { enable: true, text: 'text to insert into the text/plain portion of the message', html: 'html to insert into the text/html portion of the message', substitution_tag: 'Optional tag to replace with the open image in the body of the def message' }}
             mail['tracking_settings'] = tracking
             mailer.deliver!(mail)
-            expect(client.sent_mail['tracking_settings']).to eq(tracking)
+            expect(client.sent_mail['tracking_settings']).to eq(transform_keys(tracking, &:to_s))
           end
 
           it 'sets ganalytics' do
-            tracking = { 'ganalytics' => {'enable' => true, 'utm_source' => 'some source', 'utm_medium' => 'some medium', 'utm_term' => 'some term', 'utm_content' => 'some content', 'utm_campaign' => 'some campaign' }}
+            tracking = { ganalytics: { enable: true, utm_source: 'some source', utm_medium: 'some medium', utm_term: 'some term', utm_content: 'some content', utm_campaign: 'some campaign' }}
             mail['tracking_settings'] = tracking
             mailer.deliver!(mail)
-            expect(client.sent_mail['tracking_settings']).to eq(tracking)
+            expect(client.sent_mail['tracking_settings']).to eq(transform_keys(tracking, &:to_s))
           end
         end
 
         context 'dynamic template data' do
+          let(:template_data) do
+            { variable_1: '1', variable_2: '2' }
+          end
+          
+          before { mail['dynamic_template_data'] = template_data }
+
           it 'sets dynamic_template_data' do
-            template_data = { variable_1: '1', variable_2: '2' }
-            mail['dynamic_template_data'] = template_data
             mailer.deliver!(mail)
             expect(client.sent_mail['personalizations'].first['dynamic_template_data']).to eq(template_data)
           end
-        end
 
-        it 'sets dynamic template data and sandbox_mode' do
-          mail['mail_settings'] = '{}'
-          mailer.deliver!(mail)
-          expect(client.sent_mail['mail_settings']).to eq(nil)
+          it 'does not set unsubscribe substitutions' do
+            mailer.deliver!(mail)
+            expect(client.sent_mail['personalizations'].first).to_not have_key('substitutions')
+          end
+
+          context 'containing what looks like hash syntax' do
+            let(:template_data) do
+              { hint: 'Just use => instead of :' }
+            end
+
+            it 'does not change values inside dynamic template data' do
+              mailer.deliver!(mail)
+              expect(
+                client.sent_mail['personalizations'].first['dynamic_template_data']
+              ).to eq(template_data)
+            end
+          end
         end
 
         it 'sets dynamic template data and sandbox_mode' do
@@ -507,6 +564,262 @@ module SendGridActionMailer
           expect(content['filename']).to eq('specs.rb')
           expect(content['type']).to eq('application/x-ruby')
           expect(content['content_id'].class).to eq(String)
+          expect(content['content_id']).to include("@")
+          expect(content['content_id']).not_to include("<")
+          expect(content['content_id']).not_to include(">")
+        end
+      end
+
+      context 'with personalizations' do
+        let(:personalizations) do
+          [
+            {
+              'to' => [
+                {'email' => 'john1@example.com', 'name' => 'John 1'},
+                {'email' => 'john2@example.com', 'name' =>  'John 2'},
+              ]
+            },
+            {
+              'to' => [
+                {'email' => 'john3@example.com', 'name' => 'John 3'},
+                {'email' => 'john4@example.com'}
+              ],
+              'cc' => [
+                {'email' => 'cc@example.com'}
+              ],
+              'bcc' => [
+                {'email' => 'bcc@example.com'}
+              ],
+              'substitutions' => {
+                '%fname%' => 'Bob'
+              },
+              'subject' => 'personalized subject',
+              'send_at' => 1443636843,
+              'custom_args' => {
+                'user_id' => '343'
+              },
+              'headers' => {
+                'X-Test' => true
+              }
+            }
+          ]
+        end
+
+        before do
+          mail.to = nil
+          mail.cc = nil
+          mail.bcc = nil
+          mail['personalizations'] = personalizations
+        end
+
+        it 'sets the provided to address personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'].length).to eq(2)
+          expect(client.sent_mail['personalizations'][0]['to']).to eq(personalizations[0]['to'])
+          expect(client.sent_mail['personalizations'][1]['to']).to eq(personalizations[1]['to'])
+        end
+
+        it 'sets the provided cc address personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]).to_not have_key('cc')
+          expect(client.sent_mail['personalizations'][1]['cc']).to eq(personalizations[1]['cc'])
+        end
+
+        it 'sets the provided bcc address personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]).to_not have_key('bcc')
+          expect(client.sent_mail['personalizations'][1]['bcc']).to eq(personalizations[1]['bcc'])
+        end
+
+        it 'sets the provided subject personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]).to_not have_key('subject')
+          expect(client.sent_mail['personalizations'][1]['subject']).to eq(personalizations[1]['subject'])
+        end
+
+        it 'sets the provided headers personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]).to_not have_key('headers')
+          expect(client.sent_mail['personalizations'][1]['headers']).to eq(personalizations[1]['headers'])
+        end
+
+        it 'sets the provided custom_arg personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]).to_not have_key('custom_args')
+          expect(client.sent_mail['personalizations'][1]['custom_args']).to eq(personalizations[1]['custom_args'])
+        end
+
+        it 'sets the provided send_at personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]).to_not have_key('send_at')
+          expect(client.sent_mail['personalizations'][1]['send_at']).to eq(personalizations[1]['send_at'])
+        end
+
+        it 'sets the provided substitution personalizations' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][1]['substitutions']).to include(personalizations[1]['substitutions'])
+        end
+
+        it 'adds to the unsubscribe link substitutions' do
+          mailer.deliver!(mail)
+          expect(client.sent_mail['personalizations'][0]['substitutions']).to eq({
+            '%asm_group_unsubscribe_raw_url%' => '<%asm_group_unsubscribe_raw_url%>',
+            '%asm_global_unsubscribe_raw_url%' => '<%asm_global_unsubscribe_raw_url%>',
+            '%asm_preferences_raw_url%' => '<%asm_preferences_raw_url%>'
+          })
+          expect(client.sent_mail['personalizations'][1]['substitutions']).to include({
+            '%asm_group_unsubscribe_raw_url%' => '<%asm_group_unsubscribe_raw_url%>',
+            '%asm_global_unsubscribe_raw_url%' => '<%asm_global_unsubscribe_raw_url%>',
+            '%asm_preferences_raw_url%' => '<%asm_preferences_raw_url%>'
+          })
+        end
+
+        context 'with symbols used as keys' do
+        let(:personalizations) do
+          [
+            {
+              to: [
+                {email: 'sally1@example.com', name: 'Sally 1'},
+                {email: 'sally2@example.com', name: 'Sally 2'},
+              ]
+            }
+          ]
+          end
+
+          it 'still works' do
+            mailer.deliver!(mail)
+            expect(client.sent_mail['personalizations'].length).to eq(1)
+            expected_to = personalizations[0][:to].map { |t| transform_keys(t, &:to_s) }
+            expect(client.sent_mail['personalizations'][0]['to']).to eq(expected_to)
+          end
+        end
+
+        context 'dynamic template data passed into a personalizaiton' do
+          let(:personalization_data) do
+            {
+              'variable_1' => '1', 'variable_2' => '2'
+            }
+          end
+
+          let(:personalizations_with_dynamic_data) do
+            personalizations.tap do |p|
+              p[1]['dynamic_template_data'] = personalization_data
+            end
+          end
+
+          before do
+            mail['personalizations'] = nil
+            mail['personalizations'] = personalizations_with_dynamic_data
+          end
+
+          it 'sets the provided dynamic template data personalizations' do
+            mailer.deliver!(mail)
+            expect(client.sent_mail['personalizations'][0]).to_not have_key('dynamic_template_data')
+            expect(client.sent_mail['personalizations'][1]['dynamic_template_data']).to eq(personalization_data)
+          end
+
+          context 'dynamic template data is also set on the mail object' do
+            let(:mail_template_data) do
+              { 'variable_3' => '1', 'variable_4' => '2' }
+            end
+
+            before { mail['dynamic_template_data'] = mail_template_data.dup }
+
+            it 'sets dynamic_template_data where not also provided as a personalization' do
+              mailer.deliver!(mail)
+              expect(client.sent_mail['personalizations'][0]['dynamic_template_data']).to eq(mail_template_data)
+            end
+
+            it 'merges the template data with a personalizations dynamic data' do
+              mailer.deliver!(mail)
+              expect(client.sent_mail['personalizations'][1]['dynamic_template_data']).to eq(
+                mail_template_data.merge(personalization_data)
+              )
+            end
+          end
+        end
+
+        context 'when to is set on mail object' do
+          before { mail.to = 'test@sendgrid.com' }
+
+          it 'adds that to address as a separate personalization' do
+            mailer.deliver!(mail)
+            expect(client.sent_mail['personalizations'].length).to eq(3)
+            expect(client.sent_mail['personalizations'][0]['to']).to eq(personalizations[0]['to'])
+            expect(client.sent_mail['personalizations'][1]['to']).to eq(personalizations[1]['to'])
+            expect(client.sent_mail['personalizations'][2]['to']).to eq([{"email"=>"test@sendgrid.com"}])
+          end
+        end
+
+        context 'when cc is set on mail object' do
+          before { mail.cc = 'test@sendgrid.com' }
+
+          it 'adds that cc address as a separate personalization' do
+            mailer.deliver!(mail)
+            expect(client.sent_mail['personalizations'].length).to eq(3)
+            expect(client.sent_mail['personalizations'][0]['cc']).to eq(personalizations[0]['cc'])
+            expect(client.sent_mail['personalizations'][1]['cc']).to eq(personalizations[1]['cc'])
+            expect(client.sent_mail['personalizations'][2]['cc']).to eq([{"email"=>"test@sendgrid.com"}])
+          end
+        end
+
+        context 'when bcc is set on mail object' do
+          before { mail.bcc = 'test@sendgrid.com' }
+
+          it 'adds that bcc address as a separate personalization' do
+            mailer.deliver!(mail)
+            expect(client.sent_mail['personalizations'].length).to eq(3)
+            expect(client.sent_mail['personalizations'][0]['bcc']).to eq(personalizations[0]['bcc'])
+            expect(client.sent_mail['personalizations'][1]['bcc']).to eq(personalizations[1]['bcc'])
+            expect(client.sent_mail['personalizations'][2]['bcc']).to eq([{"email"=>"test@sendgrid.com"}])
+          end
+        end
+
+        context 'when perform_send_request false' do
+          it 'should not send and email and return json body' do
+            m = DeliveryMethod.new(perform_send_request: false, return_response: true, api_key: 'key')
+            response = m.deliver!(mail)
+            expect(response).to respond_to(:to_json)
+          end
+        end
+
+        context 'when mail_settings are present' do
+          it 'should apply mail_settings to request body' do
+            m = DeliveryMethod.new(api_key: 'key', return_response: true,  mail_settings: { sandbox_mode: {enable: true }})
+            m.deliver!(mail)
+            expect(client.sent_mail['mail_settings']).to eq("sandbox_mode" => {"enable" => true })
+          end
+
+          context 'when mail has mail_settings set' do
+            before { mail['mail_settings'] = { spam_check: { enable: true } } }
+
+            it 'should combine local mail_settings with global settings' do
+              m = DeliveryMethod.new(api_key: 'key', return_response: true,  mail_settings: { sandbox_mode: {enable: true }})
+              m.deliver!(mail)
+              expect(client.sent_mail['mail_settings']).to eq(
+                "sandbox_mode" => {"enable" => true },
+                "spam_check" => {"enable" => true },
+              )
+            end
+          end
+
+          context 'when mail contains the same setting as global settings' do
+            before do
+              mail['mail_settings'] = {
+                sandbox_mode: { enable: false },
+                spam_check: { enable: true }
+              }
+            end
+
+            it 'should apply local mail_settings on top of global settings' do
+              m = DeliveryMethod.new(api_key: 'key', return_response: true,  mail_settings: { sandbox_mode: {enable: true }})
+              m.deliver!(mail)
+              expect(client.sent_mail['mail_settings']).to eq(
+                "sandbox_mode" => {"enable" => false },
+                "spam_check" => {"enable" => true },
+              )
+            end
+          end
         end
       end
     end