+++ CHANGELOG.md
@@ -1,5 +1,73 @@
 # Change Log
 
+## 3.2.0 - 2020-02-16
+- #107 Feature/sendgrid api settings, expand settings/options to allow for other SendGrid::Api client settings
+- #106 The SendGrid API simple error handling
+- #105 Github Actions CI
+- #102 Add http_options to configurable in mailer settings
+- #101 Railtie should use ActiveSupport.on_load
+
+## 3.1.1 - 2020-11-06
+- #92 globally configured mail_settings
+- #95 add_mail_settings - support global settings
+
+## 3.1.0 - 2020-8-03
+
+### Changes
+
+- #81 Fix warnings with ruby 2.7
+- #83 Allow string type mail.to
+- #84 Content-ID values to fix inline attachments
+
+## 3.0.2 - 2020-4-20
+
+### Changes
+
+- perform_send_request setting for testing perposes
+
+## 3.0.1 - 2020-4-3
+
+### Changes
+
+- Validate error response body for empty string
+
+## 3.0.0 - 2020-3-2
+
+### Removed
+
+- Compatibility with mail gems before version 2.7
+
+## 2.6.0 - 2020-1-23
+
+### Changes
+
+- Dont send content types with dynamic templates (#69)
+
+## 2.5.0 - 2020-1-21
+
+### Changes
+
+- Add personalizations field (#60)
+
+### Fixes
+
+- Revert "Lazy load ActionMailer::Base" (#64)
+- Yank 2.4.1
+
+## 2.4.2 - 2020-1-21
+
+### Fixes
+
+- Revert "Lazy load ActionMailer::Base" (#64)
+- Yank 2.4.1
+
+## 2.4.1 - 2020-1-20
+
+### Changed
+
+- Update Travis CI settings to test on latest Ruby and mail gem version (#55)
+- Lazy load ActionMailer::Base (#57)
+
 ## 2.4.0 - 2019-07-9
 
 ### Changed